.fvecs embedding of the 100 random images
I have run a function on this to make sure the .fvecs embedding were done correctly. They match linearly with the 100 random images folder, meaning embedding 9 matches with image 9 in the image folder. The first element in the file is the length of each embedding, hint it is 512 elements. The file is written in binary following .fvecs convention.